Change BP_JVM_VERSION default from 11.* to 11
This will not change buildpack behavior. In the future we would like to change the usage of the config value to accept JVM spec versions like 8 and 11 but not semver matchers for the full openjdk vesion. This new default will work in both cases.

Also changes documentation to recommend values like 8 or 11 instead of 8.* or 11.*.

Signed-off-by: Emily Casey <ecasey@vmware.com>